Add modular beam analysis and CLI
